backend/app/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0  # JWT token validation
passlib[bcrypt]==1.7.4  # Password hashing (if needed)

# PDF Processing
PyPDF2==3.0.1
pymupdf==1.23.8

# Core Python Libraries
requests>=2.31.0
pydantic>=2.5.0
numpy>=1.24.3

# LLM Providers (Essential only)
openai>=1.3.7
google-generativeai>=0.8.5

# Adobe Hackathon Required Dependencies (Essential only)
langchain==0.3.27
langchain-openai==0.3.29
langchain-google-genai==2.0.10
langchain-community==0.3.27

# Text-to-Speech (Essential only)
google-cloud-texttospeech==2.27.0
pydub>=0.25.0

# Basic ML (will install separately if needed)
# faiss-cpu==1.7.4
# sentence-transformers>=2.2.2
# scikit-learn>=1.3.0

# Audio Processing (for TTS)
# Note: ffmpeg is required for pydub audio processing
# Install separately: apt-get install ffmpeg (Linux) or brew install ffmpeg (macOS)